requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
Django==4.2.3
djangorestframework==3.14.0
idna==3.4
lxml==4.9.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.9.1
PyPDF2==3.0.1
python-pptx==0.6.21
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
XlsxWriter==3.1.2
